refactor: use lazy component entries across platforms
