--- packages/components/src/forms/Form/index.web-only.tsx
@@ -1,2 +1,0 @@
-export { Form } from './index.shared';
-export type { IFieldErrorProps, IFieldProps, IFormProps } from './FormBase';
